Fix the Windows build and test
Tested Lua 5.1.5 mingw/VS2013 32/64 bit

+++ src/redis_hyperloglog.c
@@ -336,7 +336,7 @@ int hllPatLen(unsigned char* ele, size_t elesize, long* regp)
    *
    * This may sound like inefficient, but actually in the average case
    * there are high probabilities to find a 1 after a few iterations. */
-  hash = MurmurHash64A(ele, elesize, 0xadc83b19ULL);
+  hash = MurmurHash64A(ele, (int)elesize, 0xadc83b19ULL);
   index = hash & HLL_P_MASK; /* Register index. */
   hash |= ((uint64_t)1 << 63); /* Make sure the loop terminates. */
   bit = HLL_REGISTERS; /* First bit not used to address the register. */
